+++ cpp/example_code/transcribe-streaming/get_transcript.cpp
