+++ javav2/example_code/transcribe-streaming/src/main/java/com/amazonaws/transcribestreaming/TranscribeMedicalStreamingDemoApp.java
