+++ javav2/example_code/transcribe-streaming/src/main/java/com/amazonaws/transcribestreaming/TranscribeStreamingRetryClient.java
@@ -1,0 +1,194 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.amazonaws.transcribestreaming;
+
+import org.reactivestreams.Publisher;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
+import software.amazon.awssdk.auth.signer.EventStreamAws4Signer;
+import software.amazon.awssdk.core.client.config.SdkAdvancedClientOption;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.transcribestreaming.TranscribeStreamingAsyncClient;
+import software.amazon.awssdk.services.transcribestreaming.model.AudioStream;
+import software.amazon.awssdk.services.transcribestreaming.model.BadRequestException;
+import software.amazon.awssdk.services.transcribestreaming.model.StartStreamTranscriptionRequest;
+import software.amazon.awssdk.services.transcribestreaming.model.StartStreamTranscriptionResponseHandler;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+
+/**
+ * Build a client wrapper around the Amazon Transcribe client to retry
+ * on an exception that can be retried.
+ */
+// snippet-start:[transcribe.java-streaming-retry-client]
+public class TranscribeStreamingRetryClient {
+
+    private static final int DEFAULT_MAX_RETRIES = 10;
+    private static final int DEFAULT_MAX_SLEEP_TIME_MILLS = 100;
+    private static final Logger log = LoggerFactory.getLogger(TranscribeStreamingRetryClient.class);
+    private final TranscribeStreamingAsyncClient client;
+    List<Class<?>> nonRetriableExceptions = Arrays.asList(BadRequestException.class);
+    private int maxRetries = DEFAULT_MAX_RETRIES;
+    private int sleepTime = DEFAULT_MAX_SLEEP_TIME_MILLS;
+
+    /**
+     * Create a TranscribeStreamingRetryClient with given credential and
+     * configuration
+     */
+    public TranscribeStreamingRetryClient(AwsCredentialsProvider creds,
+            String endpoint, Region region) throws URISyntaxException {
+        this(TranscribeStreamingAsyncClient.builder()
+                .overrideConfiguration(
+                        c -> c.putAdvancedOption(
+                                SdkAdvancedClientOption.SIGNER,
+                                EventStreamAws4Signer.create()))
+                .credentialsProvider(creds)
+                .endpointOverride(new URI(endpoint))
+                .region(region)
+                .build());
+    }
+
+    /**
+     * Initiate TranscribeStreamingRetryClient with TranscribeStreamingAsyncClient
+     */
+
+    public TranscribeStreamingRetryClient(TranscribeStreamingAsyncClient client) {
+        this.client = client;
+    }
+
+    /**
+     * Get Max retries
+     */
+    public int getMaxRetries() {
+        return maxRetries;
+    }
+
+    /**
+     * Set Max retries
+     */
+    public void setMaxRetries(int maxRetries) {
+        this.maxRetries = maxRetries;
+    }
+
+    /**
+     * Get sleep time
+     */
+    public int getSleepTime() {
+        return sleepTime;
+    }
+
+    /**
+     * Set sleep time between retries
+     */
+    public void setSleepTime(int sleepTime) {
+        this.sleepTime = sleepTime;
+    }
+
+    /**
+     * Initiate a Stream Transcription with retry.
+     */
+
+    public CompletableFuture<Void> startStreamTranscription(final StartStreamTranscriptionRequest request,
+            final Publisher<AudioStream> publisher,
+            final StreamTranscriptionBehavior responseHandler) {
+
+        CompletableFuture<Void> finalFuture = new CompletableFuture<>();
+
+        recursiveStartStream(rebuildRequestWithSession(request), publisher, responseHandler, finalFuture, 0);
+
+        return finalFuture;
+    }
+
+    /**
+     * Recursively call startStreamTranscription() until the request is completed or
+     * we run out of retries.
+     *
+     */
+    private void recursiveStartStream(final StartStreamTranscriptionRequest request,
+            final Publisher<AudioStream> publisher,
+            final StreamTranscriptionBehavior responseHandler,
+            final CompletableFuture<Void> finalFuture,
+            final int retryAttempt) {
+        CompletableFuture<Void> result = client.startStreamTranscription(request, publisher,
+                getResponseHandler(responseHandler));
+        result.whenComplete((r, e) -> {
+            if (e != null) {
+                log.debug("Error occured:", e);
+
+                if (retryAttempt <= maxRetries && isExceptionRetriable(e)) {
+                    log.debug("Retriable error occurred and will be retried.");
+                    log.debug("Sleeping for sometime before retrying...");
+                    try {
+                        Thread.sleep(sleepTime);
+                    } catch (InterruptedException e1) {
+                        log.debug("Unable to sleep. Failed with exception: ", e);
+                        e1.printStackTrace();
+                    }
+                    log.debug("Making retry attempt: " + (retryAttempt + 1));
+                    recursiveStartStream(request, publisher, responseHandler, finalFuture, retryAttempt + 1);
+                } else {
+                    log.error("Encountered unretriable exception or ran out of retries. ");
+                    responseHandler.onError(e);
+                    finalFuture.completeExceptionally(e);
+                }
+            } else {
+                responseHandler.onComplete();
+                finalFuture.complete(null);
+            }
+        });
+    }
+
+    private StartStreamTranscriptionRequest rebuildRequestWithSession(StartStreamTranscriptionRequest request) {
+        return StartStreamTranscriptionRequest.builder()
+                .languageCode(request.languageCode())
+                .mediaEncoding(request.mediaEncoding())
+                .mediaSampleRateHertz(request.mediaSampleRateHertz())
+                .sessionId(UUID.randomUUID().toString())
+                .build();
+    }
+
+    /**
+     * StartStreamTranscriptionResponseHandler implements subscriber of transcript
+     * stream
+     * Output is printed to standard output
+     */
+    private StartStreamTranscriptionResponseHandler getResponseHandler(
+            StreamTranscriptionBehavior transcriptionBehavior) {
+        final StartStreamTranscriptionResponseHandler build = StartStreamTranscriptionResponseHandler.builder()
+                .onResponse(r -> {
+                    transcriptionBehavior.onResponse(r);
+                })
+                .onError(e -> {
+                    // Do nothing here. Don't close any streams that shouldn't be cleaned up yet.
+                })
+                .onComplete(() -> {
+                    // Do nothing here. Don't close any streams that shouldn't be cleaned up yet.
+                })
+
+                .subscriber(event -> transcriptionBehavior.onStream(event))
+                .build();
+        return build;
+    }
+
+    /**
+     * Check if the exception can be retried.
+     *
+     */
+    private boolean isExceptionRetriable(Throwable e) {
+        e.printStackTrace();
+
+        return nonRetriableExceptions.contains(e.getClass());
+    }
+
+    public void close() {
+        this.client.close();
+    }
+
+}
+// snippet-end:[transcribe.java-streaming-retry-client]
