+++ javav2/example_code/transcribe-streaming/src/test/java/TranscribeTest.java
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import com.amazonaws.transcribe.BidirectionalStreaming;
+import com.amazonaws.transcribestreaming.BidirectionalStreaming;
 import software.amazon.awssdk.services.transcribestreaming.TranscribeStreamingAsyncClient;
 import software.amazon.awssdk.regions.Region;
 import org.junit.jupiter.api.*;
