--- javav2/example_code/transcribe/src/main/java/com/amazonaws/transcribe/AudioStreamPublisher.java
@@ -1,110 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// snippet-start:[transcribe.java2.bidir_streaming_audiopublisher.complete]
-package com.amazonaws.transcribe;
-
-// snippet-start:[transcribe.java2.bidir_streaming_audiopublisher.import]
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.UncheckedIOException;
-import java.nio.ByteBuffer;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicLong;
-import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
-import org.reactivestreams.Subscription;
-import software.amazon.awssdk.core.SdkBytes;
-import software.amazon.awssdk.services.transcribestreaming.model.AudioEvent;
-import software.amazon.awssdk.services.transcribestreaming.model.AudioStream;
-import software.amazon.awssdk.services.transcribestreaming.model.TranscribeStreamingException;
-// snippet-end:[transcribe.java2.bidir_streaming_audiopublisher.import]
-
-// snippet-start:[transcribe.java2.bidir_streaming_audiopublisher.class]
-
-public class AudioStreamPublisher implements Publisher<AudioStream> {
-    private final InputStream inputStream;
-
-    public AudioStreamPublisher(InputStream inputStream) {
-        this.inputStream = inputStream;
-    }
-
-    @Override
-    public void subscribe(Subscriber<? super AudioStream> s) {
-        s.onSubscribe(new SubscriptionImpl(s, inputStream));
-    }
-
-    private class SubscriptionImpl implements Subscription {
-        private static final int CHUNK_SIZE_IN_BYTES = 1024 * 1;
-        private ExecutorService executor = Executors.newFixedThreadPool(1);
-        private AtomicLong demand = new AtomicLong(0);
-
-        private final Subscriber<? super AudioStream> subscriber;
-        private final InputStream inputStream;
-
-        private SubscriptionImpl(Subscriber<? super AudioStream> s, InputStream inputStream) {
-            this.subscriber = s;
-            this.inputStream = inputStream;
-        }
-
-        @Override
-        public void request(long n) {
-            if (n <= 0) {
-                subscriber.onError(new IllegalArgumentException("Demand must be positive"));
-            }
-
-            demand.getAndAdd(n);
-
-            executor.submit(() -> {
-                try {
-                    do {
-                        ByteBuffer audioBuffer = getNextEvent();
-                        if (audioBuffer.remaining() > 0) {
-                            AudioEvent audioEvent = audioEventFromBuffer(audioBuffer);
-                            subscriber.onNext(audioEvent);
-                        } else {
-                            subscriber.onComplete();
-                            break;
-                        }
-                    } while (demand.decrementAndGet() > 0);
-                } catch (TranscribeStreamingException e) {
-                    subscriber.onError(e);
-                }
-            });
-        }
-
-        @Override
-        public void cancel() {
-
-        }
-
-        private ByteBuffer getNextEvent() {
-            ByteBuffer audioBuffer;
-            byte[] audioBytes = new byte[CHUNK_SIZE_IN_BYTES];
-
-            int len = 0;
-            try {
-                len = inputStream.read(audioBytes);
-
-                if (len <= 0) {
-                    audioBuffer = ByteBuffer.allocate(0);
-                } else {
-                    audioBuffer = ByteBuffer.wrap(audioBytes, 0, len);
-                }
-            } catch (IOException e) {
-                throw new UncheckedIOException(e);
-            }
-
-            return audioBuffer;
-        }
-
-        private AudioEvent audioEventFromBuffer(ByteBuffer bb) {
-            return AudioEvent.builder()
-                    .audioChunk(SdkBytes.fromByteBuffer(bb))
-                    .build();
-        }
-    }
-}
-// snippet-end:[transcribe.java2.bidir_streaming_audiopublisher.class]
-// snippet-end:[transcribe.java2.bidir_streaming_audiopublisher.complete]
