--- javav2/example_code/transcribe/src/main/java/com/amazonaws/transcribestreaming/TranscribeStreamingDemoFile.java
@@ -1,199 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.amazonaws.transcribestreaming;
-
-// snippet-start:[transcribe.java-streaming-demo-file.import]
-import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
-import org.reactivestreams.Subscription;
-import software.amazon.awssdk.core.SdkBytes;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.transcribestreaming.TranscribeStreamingAsyncClient;
-import software.amazon.awssdk.services.transcribestreaming.model.*;
-import java.io.*;
-import java.nio.ByteBuffer;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicLong;
-// snippet-end:[transcribe.java-streaming-demo-file.import]
-
-// snippet-start:[transcribe.java-streaming-demo-file.main]
-/**
- * To run this AWS code example, ensure that you have set up your development
- * environment, including your AWS credentials.
- *
- * For information, see this documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-
-public class TranscribeStreamingDemoFile {
-    private static final Region REGION = Region.US_EAST_1;
-    private static TranscribeStreamingAsyncClient client;
-
-    public static void main(String args[]) throws ExecutionException, InterruptedException {
-
-        final String USAGE = "\n" +
-                "Usage:\n" +
-                "    <file> \n\n" +
-                "Where:\n" +
-                "    file - the location of a PCM file to transcribe. In this example, ensure the PCM file is 16 hertz (Hz). \n";
-
-        if (args.length != 1) {
-            System.out.println(USAGE);
-            System.exit(1);
-        }
-
-        String file = args[0];
-        client = TranscribeStreamingAsyncClient.builder()
-                .region(REGION)
-                .build();
-
-        CompletableFuture<Void> result = client.startStreamTranscription(getRequest(16_000),
-                new AudioStreamPublisher(getStreamFromFile(file)),
-                getResponseHandler());
-
-        result.get();
-        client.close();
-    }
-
-    private static InputStream getStreamFromFile(String file) {
-        try {
-            File inputFile = new File(file);
-            InputStream audioStream = new FileInputStream(inputFile);
-            return audioStream;
-
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static StartStreamTranscriptionRequest getRequest(Integer mediaSampleRateHertz) {
-        return StartStreamTranscriptionRequest.builder()
-                .languageCode(LanguageCode.EN_US)
-                .mediaEncoding(MediaEncoding.PCM)
-                .mediaSampleRateHertz(mediaSampleRateHertz)
-                .build();
-    }
-
-    private static StartStreamTranscriptionResponseHandler getResponseHandler() {
-        return StartStreamTranscriptionResponseHandler.builder()
-                .onResponse(r -> {
-                    System.out.println("Received Initial response");
-                })
-                .onError(e -> {
-                    System.out.println(e.getMessage());
-                    StringWriter sw = new StringWriter();
-                    e.printStackTrace(new PrintWriter(sw));
-                    System.out.println("Error Occurred: " + sw.toString());
-                })
-                .onComplete(() -> {
-                    System.out.println("=== All records stream successfully ===");
-                })
-                .subscriber(event -> {
-                    List<Result> results = ((TranscriptEvent) event).transcript().results();
-                    if (results.size() > 0) {
-                        if (!results.get(0).alternatives().get(0).transcript().isEmpty()) {
-                            System.out.println(results.get(0).alternatives().get(0).transcript());
-                        }
-                    }
-                })
-                .build();
-    }
-
-    private static class AudioStreamPublisher implements Publisher<AudioStream> {
-        private final InputStream inputStream;
-        private static Subscription currentSubscription;
-
-        private AudioStreamPublisher(InputStream inputStream) {
-            this.inputStream = inputStream;
-        }
-
-        @Override
-        public void subscribe(Subscriber<? super AudioStream> s) {
-
-            if (this.currentSubscription == null) {
-                this.currentSubscription = new SubscriptionImpl(s, inputStream);
-            } else {
-                this.currentSubscription.cancel();
-                this.currentSubscription = new SubscriptionImpl(s, inputStream);
-            }
-            s.onSubscribe(currentSubscription);
-        }
-    }
-
-    public static class SubscriptionImpl implements Subscription {
-        private static final int CHUNK_SIZE_IN_BYTES = 1024 * 1;
-        private final Subscriber<? super AudioStream> subscriber;
-        private final InputStream inputStream;
-        private ExecutorService executor = Executors.newFixedThreadPool(1);
-        private AtomicLong demand = new AtomicLong(0);
-
-        SubscriptionImpl(Subscriber<? super AudioStream> s, InputStream inputStream) {
-            this.subscriber = s;
-            this.inputStream = inputStream;
-        }
-
-        @Override
-        public void request(long n) {
-            if (n <= 0) {
-                subscriber.onError(new IllegalArgumentException("Demand must be positive"));
-            }
-
-            demand.getAndAdd(n);
-
-            executor.submit(() -> {
-                try {
-                    do {
-                        ByteBuffer audioBuffer = getNextEvent();
-                        if (audioBuffer.remaining() > 0) {
-                            AudioEvent audioEvent = audioEventFromBuffer(audioBuffer);
-                            subscriber.onNext(audioEvent);
-                        } else {
-                            subscriber.onComplete();
-                            break;
-                        }
-                    } while (demand.decrementAndGet() > 0);
-                } catch (Exception e) {
-                    subscriber.onError(e);
-                }
-            });
-        }
-
-        @Override
-        public void cancel() {
-            executor.shutdown();
-        }
-
-        private ByteBuffer getNextEvent() {
-            ByteBuffer audioBuffer = null;
-            byte[] audioBytes = new byte[CHUNK_SIZE_IN_BYTES];
-
-            int len = 0;
-            try {
-                len = inputStream.read(audioBytes);
-
-                if (len <= 0) {
-                    audioBuffer = ByteBuffer.allocate(0);
-                } else {
-                    audioBuffer = ByteBuffer.wrap(audioBytes, 0, len);
-                }
-            } catch (IOException e) {
-                throw new UncheckedIOException(e);
-            }
-
-            return audioBuffer;
-        }
-
-        private AudioEvent audioEventFromBuffer(ByteBuffer bb) {
-            return AudioEvent.builder()
-                    .audioChunk(SdkBytes.fromByteBuffer(bb))
-                    .build();
-        }
-    }
-}
-// snippet-end:[transcribe.java-streaming-demo-file.main]
